sound: wcd9335: Disable Codec Power Gating and Static Voltage Scaling
Provide an unfiltered/unrounded power readings to the DAC in order to provide a more accurate power delivery to reduce the amount of distortions due to the restricted power delivery at high volumes. This however may increase the possibility of power leakage at stand-by, in which we could sacrifice as we could afford to do so.

Signed-off-by: ederekun <sedrickvince@gmail.com>