sched/tune: Remove some unnecessary configurations
Having a negative boost is not really that beneficial in our case since we depend on task placement. Thus, move the lower limit towards 0.
There is also no chance that do_boost function will be writting the same value given our restrictions to boost and dynamic_boost. This is as long as it is prevented by the driver that uses the function.

Signed-off-by: ederekun <sedrickvince@gmail.com>