sched/dynamic_stune: Further remove unnecessary code
Since it is much faster to check an atomic state rather than consistently checking current jiffies and converting last_time into jiffies, let's instead schedule an uninterruptible timeout in the thread to delay the releasing the lock.

While we're at it let's do some adjustments to the thread and variable formats.

Signed-off-by: ederekun <sedrickvince@gmail.com>