sched/fair: Trigger iowait boost only if the task is top-app or foreground
Related article:
https://lkml.org/lkml/2020/1/23/985

Although this method is not inline with what's being discussed in the article (due to how old the kernel version is), this similarly addresses the said issue in the message regarding iowait boost being triggered by background tasks in iowait. Thus, with this commit we have added a function which checks and compares if the task's cgroup is top-app or foreground before triggering the cpufreq iowait update.
We have not used task_is_boosted because due to our dynamic stune setup, the top-app will not have consistent schedtune boost.

With this change, there are now less reasons to not enable iowait boost as this greatly reduces the amount of power consumption caused by this feature. Thus, let's remove the switch to remove unnecessary overhead.

Signed-off-by: ederekun <sedrickvince@gmail.com>