sched: Introduce SUGOV_LIMIT flag
With this flag, we can trigger sugov limit updates during runtime. In this iteration, the function affiliated with the flag will set up and down rate limits depending on another flag, SUGOV_LIMIT_ACTIVE, which indicates if the caller wants to set the active values or not.
Using the recently introduced SUGOV_LIMIT flags, let's create an update util hook in framebuffer function to request updates to sugov with the flags. Set the flags according to when the framebuffer structure is put to false state which depends on if the framebuffer stops committing or if the input structure has expired.

Signed-off-by: ederekun <sedrickvince@gmail.com>